requirements.txt
pandas
scikit-learn
langchain-openai
langchain-text-splitters
PyPDF2
faiss-cpu
pymupdf
tqdm
python-dotenv